Cleanup: do_magic, super_rm and directory_watcher are linked against libraries. nenc haves chka lurj em asum :/
