+++ super_rm/super_rm.h
@@ -3,9 +3,6 @@
 #include <string>
 #include <vector>
 
-namespace
-{
-
 constexpr size_t blockSize = 4096;
 
 std::vector<char> generateRandomVector();
@@ -15,8 +12,6 @@ std::size_t getFileSize(int fd);
 void secureDeleteFile(const std::string &filename, bool verbose);
 
 bool removeDirectory(const char* path, bool verbose);
-
-} // Unnamed namespace.
 
 #endif // SUPER_RM_H
